Pull up following revision(s) (requested by martin in ticket #846): 	sys/arch/mips/include/ansi.h: revision 1.29 	sys/arch/sh3/include/ansi.h: revision 1.16 	sys/arch/sparc64/include/ansi.h: revision 1.18 	sys/arch/m68k/include/ansi.h: revision 1.24 	sys/arch/powerpc/include/ansi.h: revision 1.30 	sys/arch/hppa/include/ansi.h: revision 1.14 	sys/arch/i386/include/ansi.h: revision 1.27 	sys/arch/alpha/include/ansi.h: revision 1.25 	sys/arch/usermode/include/ansi.h: revision 1.5 	sys/arch/sparc/include/ansi.h: revision 1.24 Make _BSD_CLOCK_T_ unsigned int so it's the same for IPL32 and LP64 environments.  We don't really have a powerpc64 native userland and the mips64 native userland is IPL32 so this shouldn't affect anything. Make clock_t unsigned Make clock_t unsigned int everywhere. Ok: matt@, mrg@

+++ sys/arch/i386/include/ansi.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ansi.h,v 1.25 2011/07/17 20:54:41 joerg Exp $	*/
+/*	$NetBSD: ansi.h,v 1.25.28.1 2015/07/16 18:27:22 riz Exp $	*/
 
 /*-
  * Copyright (c) 1990, 1993
@@ -47,7 +47,7 @@
  *	#undef	_BSD_SIZE_T_
  *	#endif
  */
-#define	_BSD_CLOCK_T_		unsigned long	/* clock() */
+#define	_BSD_CLOCK_T_		unsigned int	/* clock() */
 #define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
 #define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
 #define	_BSD_SSIZE_T_		int		/* byte count or error */
